Refactored dashboard header, split up public and standard components
