+++ frontend/components/layouts/DashboardLayout/Header/types.ts
@@ -1,0 +1,5 @@
+export interface HeaderRedirect {
+  environmentChange?: boolean;
+  projectChange?: boolean;
+  url: string;
+}
